+++ .eslintrc.js
@@ -25,43 +25,11 @@ module.exports = {
     "import/no-internal-modules": [
       "error",
       {
-        allow: ["@jobber/components/*", "gatsby-theme-docz/**", "lodash/*"]
+        allow: ["@jobber/components/*", "lodash/*"]
       }
     ]
   },
   overrides: [
-    {
-      /**
-       * Disabling default exports for a few files as docz requires
-       * some files to have default exports.
-       */
-      files: [
-        "./packages/docz-tools/src/gatsby-theme-docz/theme/index.ts",
-        "./packages/docz-tools/src/gatsby-theme-docz/components/index.ts",
-        "./packages/docz-tools/src/wrapper.tsx"
-      ],
-      rules: {
-        "import/no-default-export": "off"
-      }
-    },
-    {
-      /**
-       * Adds an alias that is only available in the docz-tools
-       * package. We do not want this available anywhere else.
-       */
-      files: ["./packages/docz-tools/**/*"],
-      settings: {
-        "import/resolver": {
-          alias: {
-            map: [
-              ...packageAliases,
-              ["~theme", "./packages/docz-tools/src/gatsby-theme-docz/theme"]
-            ],
-            extensions: [".ts", ".tsx", ".js", ".jsx", ".json"]
-          }
-        }
-      }
-    },
     {
       files: ["*.stories.mdx"],
       extends: "plugin:mdx/recommended",