+++ packages/docz-tools/src/gatsby-theme-docz/components/Actions/styles.ts
@@ -1,9 +1,0 @@
-export const actions = {
-  display: "flex",
-  paddingTop: "var(--space-base)",
-  paddingBottom: "var(--space-extravagant)",
-
-  "> *": {
-    marginRight: "small",
-  },
-} as const;